Clean up pipeline instructions
