+++ scripts/3_make_index_group.py
@@ -13,4 +13,5 @@ df2 = df.iloc[idx:]
 df1.to_csv("/Users/vanessa/dev/mila/ami-platform/make_index_1.csv")
 df2.to_csv("/Users/vanessa/dev/mila/ami-platform/make_index_2.csv")
 
-# Group the ones that appear to refer to the same todo
+# Follow the instructions in scripts/README.md about what Claude prompt
+# to use to generate groups for make_index_1 and make_index_2.
